Fix cp-web on local hyperledger networks
Commercial Paper required a lot of manual code changes in order to run
on a local hyperledger network, and some doc errors meant that users
attempting to follow this path would see errors even after implementing
the changes.  This commit automates most of those steps and fixes the
erroneous documentation.